Move test infrastructure out of src/ into test/
Move GlobalSetup.ts, StartContainers.ts, and Sandbox.ts to test/
since they are not part of the public API. Simplify tsconfig and
docgen excludes now that test helpers are outside src/.
