+++ packages/integresql/docgen.json
@@ -1,12 +1,8 @@
 {
   "$schema": "../../node_modules/@effect/docgen/schema.json",
   "exclude": [
-    "src/GlobalSetup.ts",
-    "src/StartContainers.ts",
-    "src/Sandbox.ts",
     "src/internal/**/*.ts",
-    "src/**/*.node.unit.test.ts",
-    "src/**/*.node.integration.test.ts",
+    "src/**/*.test.ts",
     "src/**/*.test-d.ts"
   ]
 }
