+++ packages/integresql/src/Examples.node.integration.test.ts
@@ -5,7 +5,7 @@ import { Effect, pipe, Redacted } from "effect"
 import { inject } from "vitest"
 import type { DatabaseConfiguration } from "./index.js"
 import { getConnection } from "./index.js"
-import { Sandbox } from "./Sandbox.ts"
+import { Sandbox } from "../test/Sandbox.js"
 import { templateIdFromFiles } from "./TemplateIdFromFilesHash.js"
 
 const dependencies = NodeContext.layer
