+++ packages/integresql/src/TemplateIdFromFilesHash.node.unit.test.ts
@@ -2,7 +2,7 @@ import { NodeContext } from "@effect/platform-node"
 import { describe, expect, it } from "@effect/vitest"
 import { Effect, Exit, pipe } from "effect"
 import path from "path"
-import { Sandbox } from "./Sandbox.ts"
+import { Sandbox } from "../test/Sandbox.js"
 import { NoMatchingFiles, templateIdFromFiles } from "./TemplateIdFromFilesHash.js"
 
 const dependencies = NodeContext.layer
